a_onsite_lectures/class10-20240114/todo_api_simple/requirements.txt
python-dotenv
SQLAlchemy
fastapi
uvicorn
pydantic
